requirements.txt
numpy
langchain
langchain_groq
langchain_community
pymongo
fastapi
uvicorn
python-multipart
python-dotenv